LVSP-61. Seperated information parsing into src/util/contact-parser.js
